// src/services/models/pageantdetails/contestant.ts
export interface Contestant {
  id?: number;
  name?: string;
  start_in_pageant?: string;
  slug?: string;
  dob?: string;
  zodiac_sign?: string;
  bio?: string;
  occupation?: string;
  college_attend?: string;
  height?: number;
  weight?: any;
  hair_color?: number;
  eye_color?: number;
  complexion?: string;
  fun_facts?: string;
  talent?: string;
  pageant_plateform?: string;
  image?: string;
  country_id?: number;
  created_by?: number;
  owner_id?: number;
  old_owner_id?: number;
  state_id?: any;
  city_id?: number;
  is_vip?: string;
  seo_focuskw?: any;
  seo_title?: any;
  seo_metadesc?: any;
  seo_linkdex?: any;
  skype_calls?: any;
  active_membership_id?: any;
  pay_profile_id?: any;
  active_membership_price?: any;
  dnd_status?: string;
  dnd_inactive_date?: any;
  collect_donation_status?: string;
  hide_message_me?: string;
  is_married?: string;
  have_kids?: string;
  status?: string;
  view_count?: number;
  search_count?: number;
  redirect_membership_date?: string;
  import_id?: any;
  first_name?: string;
  last_name?: string;
  contestant_todo_reminder_last_date?: string;
  hide_dob?: string;
  is_minor?: string;
  todo_sent_email_last_date?: any;
  public_url?: string;
  image_url?: string;
  owner?: Owner;
  pageant_award?: PageantAward[];
  contestant_hair_color?: ContestantInfo;
  contestant_eye_color?: ContestantInfo;
  contestant_zodiac?: ContestantInfo;
  contestant_weight?: ContestantInfo;
  contestant_state?: ContestantInfo;
  contestant_country?: ContestantInfo;
  contestant_height?: ContestantInfo;
  text?: string;
  address?: string;
  image_with_path?: string;
  contestant_title?: string;
  pageant_contestant_id?: number;
  pageant_id?: number;
  age_division_id?: number;
  contestant_name?: string;
  contestant_image?: string;
  contestant_id?: number;
  final_image_url?: string;
  additional_title?: string;
  additional_title_value?: string;
  title_awarded_text?: string;
  weight_id?: number;
  won?: string;
  award?: string;
  got_title?: any;
  contestant_image_url?: string;
  award_title?: string;
  contestant_status?: string;
  event_award?: Award;
  contestant_gallery?: GalleryItem[];
  email?: string;
  is_business?: string;
  business_role_id?: number;
  business_title?: string;
  image_full_url?: string;
  tagline?: string;
  longitude?: number;
  latitude?: number;
  phone?: string;
  achievement?: any;
  website?: any;
  facebook_link?: any;
  twitter_link?: any;
  instagram_link?: any;
  pinterest_link?: any;
  achievements?: any;
  detail?: any;
  you_tube_link?: any;
  google_plus_link?: any;
  custom_design?: string;
  is_featured?: string;
  provide_coaching?: string;
  hourly_rate?: any;
  rating_average?: number;
  final_rating_average?: number;
  total_user_rating?: number;
  view_phone_count?: number;
  is_verified?: string;
  is_brand?: string;
  cost_per_lead?: any;
  article_count?: any;
  disable_msg_feature?: string;
  updated_by?: any;
  role?: BusinessRole;
  record_tag?: any[];
  record_image?: any[];
  fetch_business_image?: any[];
  business_role?: BusinessRole;
  favourite?: any[];
  operating_hour?: OperatingHourMultiple;
  operating_hour_multiple?: OperatingHourMultiple[];
  social_media?: LinkedAccounts;
  is_active_advertiser?: string;
  total_votes?: number;
  votes_amount?: number;
  profile_image?: string;
  image_path?: string;
  contestant_profile?: Contestant;
  rejected_profile_claim?: any[];
  profile_claim?: any[];
  contestants?: Contestant[];
  display_contestant_name?: string;
}
export interface BusinessRole {
  id: number;
  created_by_user_id: number;
  updated_by_user_id: number;
  user_type_id: number;
  display_name: string;
  name: string;
  is_editable: string;
  is_directory: string;
  is_business: string;
  description: string;
  image: string;
  sort_order: number;
  is_for_nomination: string;
}
export interface OperatingHourMultiple {
  id: number;
  business_id: number;
  sun_to: string;
  sun_from: string;
  mon_to: string;
  mon_from: string;
  tue_to: string;
  tue_from: string;
  wed_to: string;
  wed_from: string;
  thu_to: string;
  thu_from: string;
  fri_to: string;
  fri_from: string;
  sat_to: string;
  sat_from: string;
  address: string;
  longitude: number;
  latitude: number;
  weekTiming: ExpertLocationTime[];
}
export interface ExpertLocationTime {
  time: string;
  isActive: boolean;
}
export interface PageantAward {
  id: number;
  pageant_id: number;
  age_division_id: number;
  award_id: number;
  contestant_id: number;
  pageant: Pageant;
  award: Award;
}

export interface ContestantInfo {
  id: number;
  name: string;
  slug: string;
}